+++ packages/react-querybuilder/src/utils/formatQuery/dbquery.mongodb.test.ts
@@ -39,8 +39,13 @@ afterAll(async () => {
 for (const [name, { query, expectedResult }] of Object.entries(dbTests(superUsersMongoDB))) {
   describe(name, () => {
     test('mongodb', async () => {
-      const mongoDbQuery = formatQuery(query, { format: 'mongodb' });
+      const mongoDbQuery = formatQuery(query, 'mongodb');
       const queryResult = await SuperHero.find(JSON.parse(mongoDbQuery));
+      expect(queryResult).toEqual(expectedResult.map(er => expect.objectContaining(er)));
+    });
+    test('mongodb_query', async () => {
+      const mongoDbQuery = formatQuery(query, 'mongodb_query');
+      const queryResult = await SuperHero.find(mongoDbQuery);
       expect(queryResult).toEqual(expectedResult.map(er => expect.objectContaining(er)));
     });
   });